moved cypress api support commands to a separate file
